requirements.txt
azure-iot-device
pyserial